Pulled query out of PhotosDB
