+++ osxphotos/cli/debug_dump.py
@@ -8,7 +8,7 @@ from rich import print
 
 import osxphotos
 from osxphotos._constants import _PHOTOS_4_VERSION, _UNKNOWN_PLACE
-from osxphotos.queryoptions import query_options_from_kwargs
+from osxphotos.photoquery import query_options_from_kwargs
 
 from .cli_params import (
     DB_ARGUMENT,
