+++ osxphotos/cli/sync.py
@@ -12,15 +12,15 @@ import click
 
 from osxphotos import PhotoInfo, PhotosDB, __version__
 from osxphotos.photoinfo import PhotoInfoNone
-from osxphotos.photosalbum import PhotosAlbum
-from osxphotos.photosdb.photosdb_utils import get_db_version
-from osxphotos.phototemplate import PhotoTemplate, RenderOptions
-from osxphotos.platform import assert_macos
-from osxphotos.queryoptions import (
+from osxphotos.photoquery import (
     IncompatibleQueryOptions,
     QueryOptions,
     query_options_from_kwargs,
 )
+from osxphotos.photosalbum import PhotosAlbum
+from osxphotos.photosdb.photosdb_utils import get_db_version
+from osxphotos.phototemplate import PhotoTemplate, RenderOptions
+from osxphotos.platform import assert_macos
 from osxphotos.sqlitekvstore import SQLiteKVStore
 from osxphotos.utils import pluralize
